pyproject.toml
[tool.poetry]
name = "fastapi-desafio"
version = "0.1.0"
description = ""
authors = ["CHRISTIAN S BARBOSA <32348713+djchristiandf@users.noreply.github.com>"]
readme = "README.md"

packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = { extras = ["standard"], version = "*" }
databases = { version = "*", extras = ["aiosqlite", "asyncpg"] }
pyjwt = "*"
psycopg2-binary = "*"
pydantic-settings = "*"
alembic = "*"
passlib = "*"
python-multipart = "*"

[tool.poetry.dev-dependencies]
pytest-asyncio = "*"
pytest = "*"
httpx = "*"
pytest-mock = "*"
ruff = "*"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"